# src/openapi/api.yml
openapi: '3.0.2'
info:
  title: API Title
  version: '1.0'
servers:
  - url: http://localhost:3610/v1
...
paths:
...
  /accounts/{accountId}:
    get:
      tags:
        - Accounts 📝
      security:
        - bearerAuth: []
      summary: Retrieve a specific account by ID
      description: |
        Retrieve a specific account by its unique ID.
        This endpoint allows you to fetch details of a specific account using its unique identifier (UUIDv4).
      x-eov-operation-handler: accountController
      operationId: findById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Account'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - Accounts 📝
      security:
        - bearerAuth: []
      summary: Delete a specific account by ID
      description: |
        Delete a specific account by its unique ID.
        This endpoint allows you to remove details of a specific account using its unique identifier (UUIDv4).
      x-eov-operation-handler: accountController
      operationId: deleteById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                required:
                  - statusCode
                  - message
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /accounts/{accountId}/applications/{applicationId}:
    get:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Retrieve a specific application by ID
      description: |
        Retrieve a specific application by its unique ID.
        This endpoint allows you to fetch details of a specific application using its unique identifier (UUIDv4).
      x-eov-operation-handler: applicationController
      operationId: findById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Update a specific application by ID
      description: |
        Update a specific application by its unique ID.
        This endpoint allows you to update details of a specific application using its unique identifier (UUIDv4).
      x-eov-operation-handler: applicationController
      operationId: updateById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the application.
                  example: My App
                description:
                  type: string
                  description: A brief description of the application.
                  example: This is a fantastic application
                enabled:
                  type: boolean
                  description: A flag indicating whether the application is enabled.
                  example: true
                  default: true
              additionalProperties: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Delete a specific application by ID
      description: |
        Delete a specific application by its unique ID.
        This endpoint allows you to delete details of a specific application using its unique identifier (UUIDv4).
      x-eov-operation-handler: applicationController
      operationId: deleteById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                required:
                  - statusCode
                  - message
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /accounts/{accountId}/staff/{staffId}:
    get:
      tags:
        - Staff 👨‍🏫
      security:
        - bearerAuth: []
      summary: Retrieve a specific staff member by ID
      description: |
        Retrieve a specific staff member by their unique ID.
        This endpoint allows you to fetch details of a specific staff member using their unique identifier (UUIDv4).
      x-eov-operation-handler: staffController
      operationId: findById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: staffId
          schema:
            type: string
            description: The unique identifier for the staff member (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Staff'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - Staff 👨‍🏫
      security:
        - bearerAuth: []
      summary: Delete a specific staff member by ID
      description: |
        Delete a specific staff member by their unique ID.
        This endpoint allows you to remove details of a specific staff member using their unique identifier (UUIDv4).
      x-eov-operation-handler: staffController
      operationId: deleteById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: staffId
          schema:
            type: string
            description: The unique identifier for the staff member (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                required:
                  - statusCode
                  - message
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /applications/{applicationId}/courts/{courtId}:
    get:
      tags:
        - Courts 🎾
      security:
        - bearerAuth: []
      summary: Retrieve a specific court by ID
      description: |
        Retrieve a specific court by its unique ID.
        This endpoint allows you to fetch details of a specific court using its unique identifier (UUIDv4).
      x-eov-operation-handler: courtController
      operationId: findById
      parameters:
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: courtId
          schema:
            type: string
            description: The unique identifier for the court (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Court'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      tags:
        - Courts 🎾
      security:
        - bearerAuth: []
      summary: Update a specific court by ID
      description: |
        Update a specific court by its unique ID. This endpoint allows you to update details of a specific court using its unique identifier (UUIDv4).
      x-eov-operation-handler: courtController
      operationId: updateById
      parameters:
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: courtId
          schema:
            type: string
            description: The unique identifier for the court (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the court.
                  example: 1
                sport:
                  type: string
                  description: The sport of the court.
                  example: PADEL
                  enum:
                    - PADEL
                    - TENNIS
                surface:
                  type: string
                  description: The surface of the court.
                  example: SYNTHETIC
                  enum:
                    - NATURAL_GRASS
                    - ARTIFICIAL_GRASS
                    - CLAY
                    - HARD
                    - SYNTHETIC
                enabled:
                  type: boolean
                  description: A flag indicating whether the application is enabled.
                  example: true
                  default: true
              additionalProperties: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Court'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - Courts 🎾
      security:
        - bearerAuth: []
      summary: Delete a specific court by ID
      description: |
        Delete a specific court by its unique ID.
        This endpoint allows you to delete details of a specific court using its unique identifier (UUIDv4).
      x-eov-operation-handler: courtController
      operationId: deleteById
      parameters:
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: courtId
          schema:
            type: string
            description: The unique identifier for the court (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                required:
                  - statusCode
                  - message
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  schemas:
    BadRequest:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a bad request error.
          default: 400
        message:
          type: string
          description: Descriptive message of the error.
          default: Bad Request
        errorCode:
          type: string
          enum:
            - BAD_REQUEST
          description: >
            Possible error codes:
              
            - `BAD_REQUEST`: Indicates a bad request error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    Unauthorized:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating an unauthorized error.
          default: 401
        message:
          type: string
          description: Descriptive message of the error.
          default: Unauthorized
        errorCode:
          type: string
          enum:
            - UNAUTHORIZED
          description: >
            Possible error codes:
              
            - `UNAUTHORIZED`: Indicates an unauthorized error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    Forbidden:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a forbidden error.
          default: 403
        message:
          type: string
          description: Descriptive message of the error.
          default: Forbidden
        errorCode:
          type: string
          enum:
            - FORBIDDEN
          description: >
            Possible error codes:
              
            - `FORBIDDEN`: Indicates a forbidden error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    NotFound:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a not found error.
          default: 404
        message:
          type: string
          description: Descriptive message of the error.
          default: Not Found
        errorCode:
          type: string
          enum:
            - NOT_FOUND
          description: >
            Possible error codes:
              
            - `NOT_FOUND`: Indicates a not found error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    Conflict:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a conflict error.
          default: 409
        message:
          type: string
          description: Descriptive message of the error.
          default: Conflict
        errorCode:
          type: string
          enum:
            - CONFLICT
          description: >
            Possible error codes:
              
            - `CONFLICT`: Indicates a conflict error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating an error.
          default: 500
        message:
          type: string
          description: Descriptive message of the error.
          default: Internal Server Error
        errorCode:
          type: string
          enum:
            - INTERNAL_SERVER_ERROR
          description: >
            Possible error codes:
              
            - `INTERNAL_SERVER_ERROR`: Indicates an error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
...
    Account:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the account (UUIDv4).
          example: '123e4567-e89b-12d3-a456-426614174000'
        createdAt:
          type: string
          format: date-time
          description: The time when the account was created.
          example: '2023-05-18T12:34:56Z'
        updatedAt:
          type: string
          format: date-time
          description: The time when the account was last updated.
          example: '2023-05-18T12:34:56Z'
        enabled:
          type: boolean
          description: A flag indicating whether the account is enabled.
          example: true
      required:
        - id
        - createdAt
        - updatedAt
        - enabled
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT